Resolved CID 351667 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->m_groups.size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Objects/ObjectMgr.cpp
@@ -2850,7 +2850,7 @@ void ObjectMgr::LoadGroups()
         delete result;
     }
 
-    LogDetail("ObjectMgr : %u groups loaded.", this->m_groups.size());
+    LogDetail("ObjectMgr : %u groups loaded.", static_cast<uint32_t>(this->m_groups.size()));
 }
 
 void ObjectMgr::LoadArenaTeams()
